requirements/base.txt
-e git+https://github.com/SFDO-Tooling/sfdo-template-helpers.git@v0.7.0#egg=sfdo-template-helpers
boto3==1.9.156
channels_redis==2.4.0
channels==2.2.0
dj-database-url==0.5.0
django-allauth==0.39.1
django-hashid-field==2.1.6
django-js-reverse==0.9.1
django-log-request-id==1.3.2
django-model-utils==3.1.2
django-parler==1.9.2
django-redis==4.10.0
django-rq-scheduler==1.1.3
django-rq==1.2.0  # pyup: <1.3.0
django-storages==1.7.1
Django==2.2.1
djangorestframework==3.9.4
logfmt==0.4
newrelic==4.20.0.120
Pillow==6.0.0
psycopg2-binary==2.8.2
pyopenssl==19.0.0
raven==6.10.0
redis==2.10.6  # pyup: <3.0
rq-scheduler==0.8.3  # pyup: ==0.8.3
rq==0.12  # pyup: <0.13
service_identity==18.1.0
werkzeug==0.15.4
whitenoise==4.1.2
